==> SynTwin/step6b_resolution_summarization.ipynb <==
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

def get_resolution(method):
    threshold_dict = {
        'euclidean': 250,
        'manhattan': 340,
        'cosine': 220,
        'gower': 1060
    }
    return threshold_dict[method]

import pandas as pd

# path_to_data is the folder for all input and output of SynTwin, replace path_to_data to your path
# Percolation_threshold subfolder was created in previous step
output_path = 'path_to_data/'
percolation_folder = 'Percolation_threshold'
method_list = ['euclidean','manhattan','cosine','gower']

#ran resolution on HPC for range(100,1000,10) to have a rough result 
for method in method_list:
    if method == 'gower':
        #we split the run to two part for gower, the results are saved in resolution1000_all_'+method+'.csv & resolution1000_all_'+method+'1.csv 
        df1 = pd.read_csv(output_path+percolation_folder+'/resolution1000_all_'+method+'.csv').drop(columns=['Unnamed: 0'])
        df2 = pd.read_csv(output_path+percolation_folder+'/resolution1000_all_'+method+'1.csv').drop(columns=['Unnamed: 0'])  
        df = pd.concat([df1, df2]).drop_duplicates()
    else:
        df = pd.read_csv(output_path+percolation_folder+'/resolution1000_all_'+method+'.csv').drop(columns=['Unnamed: 0'])
    df

    #for each distance method and random seed, find the resolution with the max 10members_communities
    best_df = df.sort_values(by='num_10members_communities', ascending=False).groupby(['method','random_state']).head(1)
    best_df.sort_values(by='random_state')
    
    resolution = get_resolution(method)
    best_df[best_df['resolution']==resolution]
    
    #get the most frequent resolution over the 100 replicates for each distance method
    best_count = best_df.groupby('method')['resolution'].value_counts().sort_values(ascending=False).head(5)
    best_count

# find the random state that will give the best resolution
def get_random_state(method):
    if method == 'gower':
        df1 = pd.read_csv(output_path+percolation_folder+'/resolution1000_all_'+method+'.csv').drop(columns=['Unnamed: 0'])
        df2 = pd.read_csv(output_path+percolation_folder+'/resolution1000_all_'+method+'1.csv').drop(columns=['Unnamed: 0'])  
        df = pd.concat([df1, df2]).drop_duplicates()
    else:
        df = pd.read_csv(output_path+percolation_folder+'/resolution1000_all_'+method+'.csv').drop(columns=['Unnamed: 0'])
    resolution = get_resolution(method)
    best_df = df.sort_values(by='num_10members_communities', ascending=False).groupby(['method','random_state']).head(1)
    random_seeds = best_df[(best_df['resolution']==resolution)]['random_state'] 
    return random_seeds

random_seeds1 = get_random_state('euclidean')
random_seeds2 = get_random_state('manhattan')
random_seeds3 = get_random_state('cosine')
random_seeds4 = get_random_state('gower')

pd.merge(pd.merge(pd.merge(random_seeds1, random_seeds2, on='random_state'), random_seeds3, on='random_state'), random_seeds4, on='random_state')

pd.merge(pd.merge(random_seeds1, random_seeds2, on='random_state'), random_seeds3, on='random_state')
pd.merge(pd.merge(random_seeds1, random_seeds2, on='random_state'), random_seeds4, on='random_state')
pd.merge(pd.merge(random_seeds1, random_seeds3, on='random_state'), random_seeds4, on='random_state')
pd.merge(pd.merge(random_seeds2, random_seeds3, on='random_state'), random_seeds4, on='random_state')

pd.merge(random_seeds1, random_seeds2, on='random_state')
pd.merge(random_seeds1, random_seeds3, on='random_state')
pd.merge(random_seeds1, random_seeds4, on='random_state')
pd.merge(random_seeds2, random_seeds3, on='random_state')
pd.merge(random_seeds2, random_seeds4, on='random_state')
pd.merge(random_seeds3, random_seeds4, on='random_state')
         
pd.merge(pd.merge(pd.merge(random_seeds1, random_seeds2, on='random_state'), random_seeds3, on='random_state'), random_seeds4, on='random_state')

random_seeds1
random_seeds2
random_seeds3
random_seeds4
